ended up refactoring the site structure rather than reformatting the library display, but the structure is much much better now
