+++ src/index.js
@@ -1,18 +1,36 @@
-import React from 'react';
-import ReactDOM from 'react-dom/client';
-import './index.css';
+import React, { useState } from 'react';
+import ReactDOM from 'react-dom';
 import App from './App';
+import Header from './components/Header';
+import Footer from './components/Footer';
+import AboutMe from './pages/AboutMe';
+import TestingPage from './pages/TestingPage';
+import LibraryPage from './pages/LibraryPage';
 import reportWebVitals from './reportWebVitals';
 
-const root = ReactDOM.createRoot(document.getElementById('root'));
-root.render(
+// Define the initial state and any other global states here
+const initialState = {
+  currentPageIndex: 0,
+  pages: [AboutMe, TestingPage, LibraryPage],
+};
 
-  <React.StrictMode>
-      <App />
-  </React.StrictMode>
-);
+function Index() {
+  const [state, setState] = useState(initialState);
 
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
+  const navigateToPage = (pageIndex) => {
+    setState({ ...state, currentPageIndex: pageIndex });
+  };
+
+  return (
+    <React.StrictMode>
+      <App
+        header={<Header />}
+        currentPage={state.pages[state.currentPageIndex]}
+        footer={<Footer />}
+        navigateToPage={navigateToPage}
+      />
+    </React.StrictMode>
+  );
+}
+
+ReactDOM.render(<Index />, document.getElementById('root'));
